refactor: unnecessary_cast 정리 및 clippy wrong_self_convention 경고 해결

+++ rust/overmax_app/src/ui/native_app.rs
@@ -794,7 +794,7 @@ impl NativeApp {
         let Some(ctx) = &self.session.context else {
             return false;
         };
-        let song_id = ctx.song_id as i32;
+        let song_id = ctx.song_id;
         let mode = &ctx.mode;
         let diff = &ctx.diff;
